Add Jenkins Summary widget
This widget displays a summary of specified Jenkins view. If all jobs
are green the background of the widget is green and the number of green
jobs is shown.
If one or more jobs are failing the widget background turns red and the
names of the failing jobs are displayed.